some renaming
- loose -> loss
- rollout -> environment
